Painter program
A very primitive paint program that uses circles to simulate brush strokes. There are multiple color options as well as thickness options. Because the brush strokes rely on the instantiation of circles which itself is limited by the frame rate lines are added to connect the circles. These lines only join over a certain distance beyond which the circles seperate.